Split browser and server js runtimes
