+++ packages/toolpad-app/src/toolpadDataSources/rest/runtime.tsx
@@ -1,5 +1,5 @@
 import type { Entry } from 'har-format';
-import { getBrowserRuntime } from '@mui/toolpad-core/jsRuntime';
+import { getBrowserRuntime } from '@mui/toolpad-core/jsBrowserRuntime';
 import { ExecFetchFn, RuntimeDataSource } from '../../types';
 import { FetchQuery, FetchResult } from './types';
 import { execfetch } from './shared';
