+++ packages/toolpad-app/tsconfig.json
@@ -4,6 +4,8 @@
       // TODO: remove when we move to typescript 4.7 (supports package exports)
       "@mui/toolpad-core/runtime": ["../toolpad-core/dist/runtime"],
       "@mui/toolpad-core/jsRuntime": ["../toolpad-core/dist/jsRuntime"],
+      "@mui/toolpad-core/jsBrowserRuntime": ["../toolpad-core/dist/jsBrowserRuntime"],
+      "@mui/toolpad-core/jsServerRuntime": ["../toolpad-core/dist/jsServerRuntime"],
       "@mui/toolpad-core/objectKey": ["../toolpad-core/dist/objectKey"],
       "@mui/toolpad-core/utils/*": ["../toolpad-core/dist/utils/*"]
     },
