+++ packages/toolpad-core/package.json
@@ -15,6 +15,8 @@
     "./path": "./dist/path.js",
     "./runtime": "./dist/runtime.js",
     "./jsRuntime": "./dist/jsRuntime.js",
+    "./jsServerRuntime": "./dist/jsServerRuntime.js",
+    "./jsBrowserRuntime": "./dist/jsBrowserRuntime.js",
     "./objectKey": "./dist/objectKey.js",
     "./utils/*": "./dist/utils/*.js"
   },
